+++ hospital-management/viewprescriptionrecord.php
@@ -1,16 +1,46 @@
 <?php
 include("header.php");
 include("dbconnection.php");
-if(isset($_GET['delid']))
-{
-	$sql ="DELETE FROM prescription_records WHERE prescription_record_id='$_GET[delid]'";
-	$qsql=mysqli_query($con,$sql);
-	if(mysqli_affected_rows($con) == 1)
-	{
-		echo "<script>alert('prescription record deleted successfully..');</script>";
-	}
+
+if (isset($_GET['delid'])) {
+    if (isset($_GET['confirm']) && $_GET['confirm'] === 'true') {
+        $sql = "DELETE FROM prescription_records WHERE prescription_record_id='$_GET[delid]'";
+        $qsql = mysqli_query($con, $sql);
+
+        if (mysqli_affected_rows($con) == 1) {
+            echo "<script>
+            Swal.fire({
+              title: 'Eliminado!',
+              text: 'Se ha eliminado el registro de la prescripción con éxito',
+              icon: 'success'
+            }).then(function() {
+                window.location.href = 'viewprescriptionrecords.php'; // Redirige a la página deseada después de la eliminación
+            });
+            </script>";
+        }
+    } else {
+        echo "<script>
+        Swal.fire({
+          title: '¿Estás seguro?',
+          text: 'No podrás revertirlo.',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          cancelButtonText: 'No, cancelar',
+          confirmButtonText: 'Sí, eliminar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                window.location.href = 'viewprescriptionrecords.php?delid=" . $_GET['delid'] . "&confirm=true';
+            } else {
+                window.location.href = 'viewprescriptionrecords.php'; 
+            }
+        });
+        </script>";
+    }
 }
 ?>
+
 
 <div class="wrapper col2">
   <div id="breadcrumb">
@@ -29,7 +59,7 @@ if(isset($_GET['delid']))
           <td>Unit</td>
           <td>Dosage</td>
           <td>Status</td>
-          <td>Action</td>
+          <td>Acxxxtion</td>
         </tr>
          <?php
 		$sql ="SELECT * FROM prescription_records";
