+++ hospital-management/viewtreatmentrecord.php
@@ -1,16 +1,46 @@
 <?php
 include("adformheader.php");
 include("dbconnection.php");
-if(isset($_GET['delid']))
-{
-	$sql ="DELETE FROM treatment_records WHERE appointmentid='$_GET[delid]'";
-	$qsql=mysqli_query($con,$sql);
-	if(mysqli_affected_rows($con) == 1)
-	{
-		echo "<script>alert('appointment record deleted successfully..');</script>";
-	}
+
+if (isset($_GET['delid'])) {
+    if (isset($_GET['confirm']) && $_GET['confirm'] === 'true') {
+        $sql = "DELETE FROM treatment_records WHERE appointmentid='$_GET[delid]'";
+        $qsql = mysqli_query($con, $sql);
+
+        if (mysqli_affected_rows($con) == 1) {
+            echo "<script>
+            Swal.fire({
+              title: 'Eliminado!',
+              text: 'Se ha eliminado el registro de la cita con éxito',
+              icon: 'success'
+            }).then(function() {
+                window.location.href = 'viewtreatmentrecords.php'; // Redirige a la página deseada después de la eliminación
+            });
+            </script>";
+        }
+    } else {
+        echo "<script>
+        Swal.fire({
+          title: '¿Estás seguro?',
+          text: 'No podrás revertirlo.',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          cancelButtonText: 'No, cancelar',
+          confirmButtonText: 'Sí, eliminar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                window.location.href = 'viewtreatmentrecords.php?delid=" . $_GET['delid'] . "&confirm=true';
+            } else {
+                window.location.href = 'viewtreatmentrecords.php'; 
+            }
+        });
+        </script>";
+    }
 }
 ?>
+
 
 <div class="container-fluid">
   <div class="block-header">
